Connection between Booking and User

+++ fleximove-backend/user-context-service/src/main/java/de/fhdo/spring/user/context/bootstrap/DummyDataBootstrap.java
@@ -25,6 +25,9 @@ public class DummyDataBootstrap implements ApplicationListener<ContextRefreshedE
     }
 
     private void initData() {
+    	long count = userRepository.count();
+    	System.out.println("initData called — users in DB: " + count);
+    	if (count > 0) return;
         userRepository.deleteAll();
     	 // Adresse erstellen
         Adress address = new Adress("Musterstraße", "Deutschland", "Berlin", "15", "10115");
